updates
git-svn-id: svn+ssh://svn.cern.ch/reps/atlasphys/Physics/Higgs/HSG4/software/common/externaltools/trunk@40760 883ba7d9-fdd0-4202-9341-49aa55999ad8

+++ Makefile
@@ -5,7 +5,7 @@ PREFIX ?= externaltools
 all: config build install
 
 clean-pyc:
-	find $(PREFIX) -name "*.pyc" | xargs rm -f
+	test -d $(PREFIX) || find $(PREFIX) -name "*.pyc" | xargs rm -f
 
 clean:
 	./waf distclean
